pyproject.toml
[tool.poetry]
name = "appeals"
version = "0.1.0"
description = ""
authors = ["MrMihen13 <konstantin.merezhnikov13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
gunicorn = "^21.1.0"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
djoser = "^2.2.0"
drf-yasg = "^1.21.6"
djangorestframework-simplejwt = "^5.2.2"
markdown = "^3.4.3"
django-cors-headers = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
